fixed a few things with party endpoint

+++ app/api/party/routes.py
@@ -4,6 +4,7 @@ from api.account.auth import login_required
 from api.game.gametype import str_to_type_socket,game_type
 from models.db import db
 from models.party import Party, PartyMember
+from models.user import User
 from utils import return_400_on_error
 
 from fsocket import socketio
@@ -15,6 +16,9 @@ party_bp = Blueprint("party_bp", __name__)
 def create_party(user):
     party = Party(host_id=user.id)
     db.session.add(party)
+    db.session.flush()
+    member = PartyMember(user_id=user.id, party_id=party.id)
+    db.session.add(member)
     db.session.commit()
     
     return jsonify({"code": party.code}), 200
@@ -69,7 +73,7 @@ def join_party(user):
     db.session.add(member)
     db.session.commit()
     
-    socketio.emit("new_user", {"user": user.to_json()}, namespace="/socket/party", room=code)
+    socketio.emit("add_user", {"user": user.to_json()}, namespace="/socket/party", room=code)
     
     return jsonify({"message": "joined party"}), 200
 
@@ -85,6 +89,15 @@ def start_game(user):
         return jsonify({"error": "No session yet"}), 404
     
     return return_400_on_error(game_type[session.type].join, data, user, session)
+
+@party_bp.route("host", methods=["GET"])
+@login_required
+def is_host(user):
+    data = request.args
+    code = data.get("code")
+    
+    party = Party.query.filter_by(code=code).first_or_404("Cannot find party")
+    return jsonify({"is_host":party.host_id==user.id}), 200
 
 @party_bp.route("/game/state", methods=["GET"])
 @login_required
@@ -114,24 +127,31 @@ def get_users(user):
     
     members = [member.user.to_json() for member in party.members]
     
-    return jsonify({"members": members}), 200
+    return jsonify({"members": members,"host":party.host.username,"this":user.username}), 200
 
 @party_bp.route("/users/remove", methods=["POST"])
 @login_required
 def remove_user(user):
     data = request.get_json()
     code = data.get("code")
-    user_id = data.get("user_id")
+    username = data.get("username")
+    reason = data.get("reason") if data.get("reason") else "removed from party"
     
+    remove_user = User.query.filter_by(username=username).first_or_404("Cannot find user")
     party = Party.query.filter_by(code=code).first_or_404("Cannot find party")
     
     if party.host_id != user.id:
         return jsonify({"error": "You are not the host of this party"}), 403
     
-    member = PartyMember.query.filter_by(party_id=party.id, user_id=user_id).first_or_404("Cannot find member")
+    if remove_user.user_id == party.host_id:
+        return jsonify({"error": "You cannot remove the host"}), 403
     
+    member = PartyMember.query.filter_by(party_id=party.id, user_id=remove_user.id).first_or_404("Cannot find member")
     db.session.delete(member)
     db.session.commit()
+    
+    socketio.emit("leave",{"reason":reason},namespace="/socket/party", room=f"{member.user_id}_{code}")
+    socketio.emit("remove_user", {"username": member.user.username}, namespace="/socket/party", room=code)
     
     return jsonify({"message": "removed user"}), 200
 
@@ -142,9 +162,12 @@ def leave_party(user):
     code = data.get("code")
     
     party = Party.query.filter_by(code=code).first_or_404("Cannot find party")
-    member = PartyMember.query.filter_by(party_id=party,user_id=user.id).first_or_404("Cannot find member")
+    member = PartyMember.query.filter_by(party_id=party.id,user_id=user.id).first_or_404("Cannot find member")
     db.session.delete(member)
     db.session.commit()
+    
+    socketio.emit("leave",namespace="/socket/party", room=f"{member.user_id}_{code}")
+    socketio.emit("remove_user", {"user": user.username}, namespace="/socket/party", room=code)
     
     return jsonify({"message": "left party"}), 200
 
@@ -161,6 +184,8 @@ def delete_party(user):
     
     db.session.delete(party)
     db.session.commit()
+    
+    socketio.emit("leave", {"reason": "party disbanded"}, namespace="/socket/party", room=code)
     
     return jsonify({"message": "deleted party"}), 200
